// scripts/src/libs/alkanes/types.ts
export interface AlkaneId {
  block: string;
  tx: string;
}

export interface Rune {
  rune: {
    id: { block: string; tx: string };
    name: string;
    spacedName: string;
    divisibility: number;
    spacers: number;
    symbol: string;
  };
  balance: string;
}

export interface ProtoRunesToken {
  id: AlkaneId;
  name: string;
  symbol: string;
}

export interface AlkanesOutpoint {
  runes: Rune[];
  outpoint: { txid: string; vout: number };
  output: { value: string; script: string };
  txindex: number;
  height: number;
}

export interface AlkanesResponse {
  outpoints: AlkanesOutpoint[];
  balanceSheet: [];
}

export interface AlkaneSimulateRequest {
  alkanes: any[];
  transaction: string;
  block: string;
  height: string;
  txindex: number;
  target: {
    block: string;
    tx: string;
  };
  inputs: string[];
  pointer: number;
  refundPointer: number;
  vout: number;
}

export interface AlkaneToken {
  name: string;
  symbol: string;
  totalSupply: number;
  cap: number;
  minted: number;
  mintActive: boolean;
  percentageMinted: number;
  mintAmount: number;
}

export type AlkanesByAddressResponse = {
  outpoints: AlkanesOutpoint[];
};

export interface AlkanesPayload {
  body: Uint8Array;
  cursed: boolean;
  tags: { contentType: string };
}